refactor(memory): share number formatters
